Restructured node helper
